+++ front-end/src/types/api.types.ts
@@ -5,7 +5,7 @@ export interface ApiResponse<T> {
     currentPage: number;
     pageSize: number;
 }
-export interface Pagination {
+export interface PaginationProps {
     totalPages: number;
     currentPage: number;
     pageSize: number;
